+++ src/FloorOverlay.tsx
@@ -11,11 +11,12 @@ import Overlay from "./Overlay";
 import OverlaySearchBar from "./OverlaySearchBar";
 import { css, cx, mixins, theme } from "./style";
 import { createSearchMatcher, uiText } from "./util";
+import { FloorData } from "./web-sdk";
 
 export interface FloorOverlayProps {
   toggleFloorOverlay: (options: { open: boolean }) => void;
   currentFloorID: string;
-  floors: Record<string, any>[];
+  floors: FloorData[];
   selectFloorByID: (floorID: string) => void;
 }