+++ src/TagListOverlay.tsx
@@ -13,14 +13,14 @@ import Overlay from "./Overlay";
 import OverlaySearchBar from "./OverlaySearchBar";
 import { css, mixins, theme } from "./style";
 import { createSearchMatcher, getTagLabels, uiText } from "./util";
-import { API, CreateMapOptions } from "./web-sdk";
+import { API, CreateMapOptions, FloorData, TagData } from "./web-sdk";
 
 export interface TagListOverlayProps {
-  onMarkerClick: (marker: Record<string, any>) => void;
+  onTagClick: (tag: TagData) => void;
   loading: boolean;
-  tags: Record<string, any>[];
+  tags: TagData[];
   showControlTags: boolean;
-  floors: Record<string, any>[];
+  floors: FloorData[];
   tagOptions: CreateMapOptions["tags"];
   updateMap: (options: CreateMapOptions) => void;
   api: API;
@@ -47,7 +47,7 @@ class TagListOverlay extends Component<TagListOverlayProps> {
       tagOptions = {},
       tags,
       loading,
-      onMarkerClick,
+      onTagClick,
       toggleTagListOverlay
     } = this.props;
     const { searchFilter } = this.state;
@@ -146,7 +146,7 @@ class TagListOverlay extends Component<TagListOverlayProps> {
                           floorID: tag.map_id,
                           tags: { ...tagOptions, filter: () => true }
                         });
-                        onMarkerClick(tag);
+                        onTagClick(tag);
                         toggleTagListOverlay({ open: false });
                       }}
                     >