+++ src/web-sdk.tsx
@@ -56,7 +56,7 @@ import { h, render } from "preact";
 import ReconnectingWebSocket from "reconnecting-websocket";
 import placemarkIconGeneric from "../files/placemarks/generic.svg";
 import { sendAnalyticsCodeEvent } from "./analytics";
-import MapComponent from "./Map";
+import MapComponent from "./MapComponent";
 import {
   asyncClientCall,
   logDeprecated,
@@ -119,9 +119,9 @@ if (document.readyState === "complete") {
 }
 
 /** @internal */
-type APIContext = {
+interface APIContext {
   api?: API;
-};
+}
 
 /** @internal */
 const context: APIContext = {
@@ -201,29 +201,29 @@ export function init(options: { api: API }): void {
   context.api = options.api;
 }
 
-export type CreateMapTagsOptions = {
+export interface CreateMapTagsOptions {
   /** Should we show control tags? Defaults to false. */
   showControlTags?: boolean;
   /**
    * Filter function used to hide tags. Return false to hide a tag. Defaults
    * to `() => true`.
    */
-  filter?: (tag: Record<string, any>) => boolean;
+  filter?: (tag: TagData) => boolean;
   /** Disable clicking tags when true. Defaults to false. */
   disabled?: boolean;
-};
+}
 
-export type CreateMapPlacemarksOptions = {
+export interface CreateMapPlacemarksOptions {
   /** Should we show hidden placemarks? Defaults to false. */
   showHiddenPlacemarks?: boolean;
   /**
    * Filter function used to hide placemarks. Return false to hide a
    * placemark. Defaults to `() => true`.
    */
-  filter?: (placemark: Record<string, any>) => boolean;
+  filter?: (placemark: PlacemarkData) => boolean;
   /** Disable clicking placemarks when true. Defaults to false. */
   disabled?: boolean;
-};
+}
 
 /**
  * Object describing a polygon overlay drawn on the map
@@ -276,7 +276,7 @@ export type CustomAnnotation = CustomAnnotationPoint;
 /**
  * Options passed to [[createMap]].
  */
-export type CreateMapOptions = {
+export interface CreateMapOptions {
   /** See [[restrictedPanZoom]]. */
   shouldMapPanZoom?: (event: TouchEvent | WheelEvent | MouseEvent) => boolean;
   /** Width of the map (e.g. "100%" or "300px"). */
@@ -303,19 +303,24 @@ export type CreateMapOptions = {
   placemarks?: CreateMapPlacemarksOptions;
   /** An array of custom overlays to draw on the map. */
   overlays?: CustomOverlay[];
+  /** An array of custom annotations to draw on the map. */
+  annotations?: CustomAnnotation[];
+  /**
+   * Called when the user clicks on the map. This is mostly useful as a way of
+   * knowing that the user has potentially unfocused a tag or placemark. This is
+   * NOT called when the user pans or zooms the map.
+   */
+  onMapClick?: () => void;
   /**
    * Called when a tag is clicked. Use `event.preventDefault()` to prevent the
    * default dialog from appearing.
    */
-  onTagClick?: (tag: Record<string, any>, event: MeridianEvent) => void;
+  onTagClick?: (tag: TagData, event: MeridianEvent) => void;
   /**
    * Called when a placemark is clicked. Use `event.preventDefault()` to prevent
    * the default dialog from appearing.
    */
-  onPlacemarkClick?: (
-    placemark: Record<string, any>,
-    event: MeridianEvent
-  ) => void;
+  onPlacemarkClick?: (placemark: PlacemarkData, event: MeridianEvent) => void;
   /**
    * Called when tags on the current floor are updated. `allTags` is every tag
    * on the current floor, even ones not shown on the map. `filteredTags` is
@@ -323,8 +328,8 @@ export type CreateMapOptions = {
    * `filter`).
    */
   onTagsUpdate?: (tags: {
-    allTags: Record<string, any>[];
-    filteredTags: Record<string, any>[];
+    allTags: TagData[];
+    filteredTags: TagData[];
   }) => void;
   /**
    * Called when tags on the current floor are updated. `allPlacemarks` is every
@@ -333,38 +338,38 @@ export type CreateMapOptions = {
    * `showHiddenPlacemarks` and `filter`).
    */
   onPlacemarksUpdate?: (placemarks: {
-    allPlacemarks: Record<string, any>[];
-    filteredPlacemarks: Record<string, any>[];
+    allPlacemarks: PlacemarkData[];
+    filteredPlacemarks: PlacemarkData[];
   }) => void;
   /**
    * Called with an array of floors after the floors list is updated.
    */
-  onFloorsUpdate?: (floors: Record<string, any>[]) => void;
+  onFloorsUpdate?: (floors: FloorData[]) => void;
   /**
    * Called with a floor object when the floor is changed.
    */
-  onFloorChange?: (floor: Record<string, any>) => void;
+  onFloorChange?: (floor: FloorData) => void;
   /**
    * Called when the map has been destroyed, either by manually calling
    * [[destroy]] or by being automatically destroyed when its DOM is tampered
    * with.
    */
   onDestroy?: () => void;
-};
+}
 
 /**
  * MeridanSDK specific event object, used to `preventDefault` when overriding a
  * handler.
  */
-export type MeridianEvent = {
+export interface MeridianEvent {
   preventDefault: () => void;
-};
+}
 
 /**
  * Returned from [[createMap]], this object allows you to manipulate a map that
  * has already been created in the page.
  */
-export type MeridianMap = {
+export interface MeridianMap {
   /**
    * Remove the Meridian Map from the DOM and clean up all ongoing network
    * connections.
@@ -388,7 +393,7 @@ export type MeridianMap = {
    * Zoom to a given x, y coordinate and scale to a given zoom factor.
    */
   zoomToPoint: (options: { x: number; y: number; scale: number }) => void;
-};
+}
 
 /**
  * Creates and returns a map object mounted at the given HTML element. If you
@@ -447,9 +452,14 @@ export function createMap(
     { internalUpdate = true } = {}
   ) => {
     options = { ...options, ...updatedOptions };
+    const api = context.api || options.api;
+    if (!api) {
+      requiredParam("createMap", "options.api");
+      throw new Error("couldn't create MeridianMap");
+    }
     domRef = render(
       <MapComponent
-        api={context.api}
+        api={api}
         {...options}
         update={_update}
         ref={setMapRef}
@@ -469,9 +479,14 @@ export function createMap(
       internalUpdate
     });
   };
+  const api = context.api || options.api;
+  if (!api) {
+    requiredParam("createMap", "options.api");
+    throw new Error("couldn't create MeridianMap");
+  }
   let domRef: HTMLElement = render(
     <MapComponent
-      api={context.api}
+      api={api}
       {...options}
       update={_update}
       ref={setMapRef}
@@ -550,11 +565,11 @@ export interface OpenStreamOptions {
   /** Meridian floor ID */
   floorID: string;
   /** Called with ALL tags on first load */
-  onInitialTags?: (tags: Record<string, any>[]) => void;
+  onInitialTags?: (tags: TagData[]) => void;
   /** Called when a tag exits the floor */
-  onTagLeave?: (tag: Record<string, any>) => void;
+  onTagLeave?: (tag: TagData) => void;
   /** Called when a tag location updates */
-  onTagUpdate?: (tag: Record<string, any>) => void;
+  onTagUpdate?: (tag: TagData) => void;
   /** Called when an error happens */
   onException?: (error: Error) => void;
   /** Called when the stream closes */
@@ -654,7 +669,7 @@ export class API {
   async fetchTagsByFloor(
     locationID: string,
     floorID: string
-  ): Promise<Record<string, any>[]> {
+  ): Promise<TagData[]> {
     if (!locationID) {
       requiredParam("fetchTagsByFloor", "locationID");
     }
@@ -671,9 +686,7 @@ export class API {
   /**
    * [async] Returns an array of all tags at the specified location
    */
-  async fetchTagsByLocation(
-    locationID: string
-  ): Promise<Record<string, any>[]> {
+  async fetchTagsByLocation(locationID: string): Promise<TagData[]> {
     if (!locationID) {
       requiredParam("fetchTagsByLocation", "locationID");
     }
@@ -690,7 +703,7 @@ export class API {
   async fetchPlacemarksByFloor(
     locationID: string,
     floorID: string
-  ): Promise<Record<string, any>[]> {
+  ): Promise<FloorData[]> {
     if (!locationID) {
       requiredParam("fetchPlacemarksByFloor", "locationID");
     }
@@ -706,9 +719,7 @@ export class API {
   /**
    * [async] Returns an array of all floors at the specified location
    */
-  async fetchFloorsByLocation(
-    locationID: string
-  ): Promise<Record<string, any>[]> {
+  async fetchFloorsByLocation(locationID: string): Promise<LocationData[]> {
     if (!locationID) {
       requiredParam("fetchFloorsByLocation", "locationID");
     }
@@ -924,7 +935,10 @@ export type EnvOptions =
  * });
  * ```
  */
-export type APIOptions = { environment?: EnvOptions; token: string };
+export interface APIOptions {
+  environment?: EnvOptions;
+  token: string;
+}
 
 /**
  * An open tag stream that can be closed. Returned by [[API.openStream]].
@@ -941,6 +955,34 @@ export type APIOptions = { environment?: EnvOptions; token: string };
  * stream.close();
  * ```
  */
-export type Stream = {
+export interface Stream {
   close: () => void;
-};
+}
+
+/** Meridian Tag data */
+export interface TagData {
+  [key: string]: any;
+  /** Tag MAC address (uppercase, no punctuation) */
+  mac: string;
+}
+
+/** Meridian Placemark data */
+export interface PlacemarkData {
+  [key: string]: any;
+  /** Placemark ID */
+  id: string;
+}
+
+/** Meridian Floor data */
+export interface FloorData {
+  [key: string]: any;
+  /** Floor ID */
+  id: string;
+}
+
+/** Meridian Location data */
+export interface LocationData {
+  [key: string]: any;
+  /** Location ID */
+  id: string;
+}