+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.13.2] - 2021-02-24
+## Unreleased
+
+### Changed
 
 - Upgraded the map SVG for the examples
+
+### Added
+
+- `onMapClick()` option which is called when the map is clicked
+- `TagData` interface (replaces `Record<string, any>`)
+- `PlacemarkData` interface (replaces `Record<string, any>`)
+- `FloorData` interface (replaces `Record<string, any>`)
+- `LocationData` interface (replaces `Record<string, any>`)
+
+### Removed
+
+- Removed undocumented and unsupported option `onMarkerClick`
 
 ## [0.13.1] - 2021-02-19